created Order Table and joined with user and address.

+++ src/main/java/com/example/hibarnet_testing/repositories/AddressRepo.java
@@ -3,5 +3,5 @@ package com.example.hibarnet_testing.repositories;
 import com.example.hibarnet_testing.domain.Address;
 import org.springframework.data.repository.CrudRepository;
 
-public interface AddressRepo extends CrudRepository<Address,Integer> {
+public interface AddressRepo extends CrudRepository<Address,Long> {
 }
